Create a page to allow users to view their bookings
Created a new page and component to allow clients to view the appointments
they have booked and allows employees to view appointments that have been
booked with them.
Achieved this by creating a getFilteredAppointments function in my controllers
file and using that function in our endpoint to retrieve filtered appointments
based on the authenticated user's information.